herb_deploy/yolo_api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
ultralytics
opencv-python-headless
